Move YAML parser into config.py, add class for configuring default reporters
