+++ blues/simulation.py
@@ -20,338 +20,6 @@ from simtk.openmm import app
 
 logger = logging.getLogger(__name__)
 
-def startup(yaml_config):
-    """
-    Function that will parse the YAML configuration file for setup and running
-    BLUES simulations.
-
-    Parameters
-    ----------
-    yaml_config : filepath to YAML file (or JSON)
-    """
-
-    #Default parmed units.
-    default_units = {'nonbondedCutoff':unit.angstroms,
-                    'switchDistance':unit.angstroms,
-                    'implicitSolventKappa':unit.angstroms,
-                    'freeze_distance': unit.angstroms,
-
-                    'implicitSolventSaltConc':unit.mole/unit.liters,
-                    'temperature':unit.kelvins,
-
-                    'hydrogenMass':unit.daltons,
-
-                    'dt':unit.picoseconds,
-                    'friction':1/unit.picoseconds,
-
-                    'pressure': unit.atmospheres,
-
-                    'weight': unit.kilocalories_per_mole/unit.angstroms**2,
-
-                    }
-
-    #System related parameters that require import from the simtk.openmm.app namesapce
-    valid_apps = {
-                'nonbondedMethod' : ['NoCutoff', 'CutoffNonPeriodic',
-                                    'CutoffPeriodic', 'PME', 'Ewald'],
-                'constraints' : [None, 'HBonds', 'HAngles', 'AllBonds'],
-                'implicitSolvent' : ['HCT', 'OBC1', 'OBC2', 'GBn', 'GBn2']
-                }
-    #     scalar_options = ['soluteDielectric', 'solvent', 'ewaldErrorTolerance']
-    #     bool_options = ['rigidWater', 'useSASA', 'removeCMMotion', 'flexibleConstraints', 'verbose',
-    #                     'splitDihedrals']
-    #
-
-    def load_yaml(yaml_config):
-        """
-        Function that reads the YAML configuration file and parameters are
-        returned as a dict.
-        """
-        #Parse input parameters from YAML
-        with open(yaml_config, 'r') as stream:
-            try:
-                config = (yaml.load(stream))
-            except Exception as exc:
-                print (exc)
-        return config
-
-    def check_amber_selection(structure, selection, logger):
-        """
-        Given a AmberMask selection (str) for selecting atoms to freeze or restrain,
-        check if it will actually select atoms. If the selection produces None,
-        suggest valid residues or atoms.
-
-        Parameters
-        ----------
-        structure: parmed.Structure object
-        selection: str (Amber syntax), atoms to restrain/freeze during simulation.
-        logger: logging.Logger object, records information
-        """
-
-        mask_idx = []
-        mask = parmed.amber.AmberMask(structure, str(selection))
-        mask_idx = [i for i in mask.Selected()]
-        if not mask_idx:
-            if ':' in selection:
-                res_set = set(residue.name for residue in structure.residues)
-                logger.error("'{}' was not a valid Amber selection. \n\tValid residue names: {}".format(selection, res_set))
-            elif '@' in selection:
-                atom_set = set(atom.name for atom in structure.atoms)
-                logger.error("'{}' was not a valid Amber selection. Valid atoms: {}".format(selection, atom_set))
-            sys.exit(1)
-
-    def load_Structure(filename, restart=None, logger=None, *args, **kwargs):
-        """
-        Load the input/reference files (.prmtop, .inpcrd) into a parmed.Structure. If a `restart` (.rst7)
-        file is given, overwrite the reference positions, velocities, and box vectors on the Structure.
-
-        Parameters
-        -----------
-        filename: str, filepath to input (.prmtop)
-        restart: str, file path to Amber restart file (.rst7)
-        logger: logging.Logger object, records information
-
-        Args
-        ----
-
-        Kwargs
-        ------
-
-        """
-        structure = parmed.load_file(filename, *args, **kwargs)
-        if restart:
-            logger.info('Restarting simulation from {}'.format(restart))
-            restart = parmed.amber.Rst7(restart)
-            structure.positions = restart.positions
-            structure.velocities = restart.velocities
-            structure.box = restart.box
-
-        return structure
-
-    def parse_unit_quantity(unit_quantity_str):
-        """
-        Utility for parsing parameters from the YAML file that require units.
-        Takes a str, i.e. '3.024 * daltons' and returns as a simtk.unit.Quantity
-        `unit.Quantity(3.024, unit=dalton)``
-        """
-        value, u = unit_quantity_str.replace(' ', '').split('*')
-        if '/' in u:
-            u = u.split('/')
-            return unit.Quantity(float(value), eval('%s/unit.%s' % (u[0],u[1])))
-        return unit.Quantity(float(value), eval('unit.%s' % u))
-
-    def set_Output(config):
-        """
-        Parses/updates the config (dict) with the given path for storing output files.
-
-        """
-        #Set file paths
-        if 'output_dir' in config.keys():
-            output_dir = config['output_dir']
-        else:
-            output_dir = '.'
-        outfname = os.path.join(output_dir, config['outfname'])
-        config['outfname'] = outfname
-        config['simulation']['outfname'] = outfname
-        return config
-
-    def set_Logger(config):
-        """
-        Initializes the logging.Logger modules and parses/updates the
-        config (dict) with the logger_level and the file path to store the .log file
-
-        """
-        #Initialize root Logger module
-        level = config['logger_level'].upper()
-        outfname = config['outfname']
-        if level == 'DEBUG':
-            #Add verbosity if logging is set to DEBUG
-            config['verbose'] = True
-            config['system']['verbose'] = True
-            config['simulation']['verbose'] = True
-        else:
-            config['verbose'] = False
-            config['system']['verbose'] = False
-            config['simulation']['verbose'] = False
-        logger_level = eval("logging.%s" % level)
-        logger = reporters.init_logger(logging.getLogger(), logger_level, outfname)
-        config['Logger'] = logger
-
-        return config
-
-    def set_Units(config):
-        """
-        Parses/updates the config (dict) values with parameters that should have
-        units on them. If no unit is provided, the default units are assumed.
-
-            Distances: unit.angstroms
-            Temperature: unit.kelvins
-            Masses: unit.daltons
-            Time: unit.picoseconds
-            Pressure: unit.atmospheres
-            Force:  unit.kilocalories_per_mole/unit.angstroms**2
-
-        """
-        #Loop over parameters which require units
-        for param, unit_type in default_units.items():
-
-            #Check each nested subset of parameters
-            for setup_keys in ['system', 'simulation', 'freeze', 'restraints']:
-                #If the parameter requires units, cheeck if provided by user
-                try:
-                    if param in config[setup_keys]:
-                        user_input = config[setup_keys][param]
-
-                        if '*' in str(user_input):
-                            config[setup_keys][param] =  parse_unit_quantity(user_input)
-
-                        #If not provided, set default units
-                        else:
-                            config['Logger'].warn("Units for '{} = {}' not specified. Setting units to '{}'".format(param, user_input, unit_type))
-                            config[setup_keys][param] = user_input*unit_type
-
-                except:
-                    pass
-
-        return config
-
-    def set_Apps(config):
-        """
-        Check system parameters which require loading from the simtk.openmm.app namespace
-
-
-        nonbondedMethod : ['NoCutoff', 'CutoffNonPeriodic', 'CutoffPeriodic', 'PME', 'Ewald'],
-        constraints : [None, 'HBonds', 'HAngles', 'AllBonds'],
-        implicitSolvent : ['HCT', 'OBC1', 'OBC2', 'GBn', 'GBn2']
-
-        """
-        for method, app_type in valid_apps.items():
-            if method in config['system']:
-                user_input = config['system'][method]
-                try:
-                    config['system'][method] = eval("app.%s" % user_input)
-                except:
-                    config['Logger'].exception("'{}' was not a valid option for '{}'. Valid options: {}".format(user_input, method, app_type))
-        return config
-
-    def set_Reporters(config):
-        """
-        Initializes the openmm.Reporters for the simulations
-        """
-        outfname = config['outfname']
-
-        #Configure the MD simulation reporters
-        totalStepsMD = config['simulation']['nIter']*config['simulation']['nstepsMD']
-        md_report_interval = config['simulation']['reporters']['md']['reporter_interval']
-        md_trajectory_interval = config['simulation']['reporters']['md']['trajectory_interval']
-
-        if 'frame_indices' in config['simulation']['reporters']['md']:
-            frame_indices = config['simulation']['reporters']['md']['frame_indices']
-        else:
-            frame_indices = None
-        md_reporters = reporters.getReporters(totalStepsMD, outfname, progress=False, **config['simulation']['reporters']['md'])
-        md_progress_reporter = reporters.BLUESStateDataReporter(config['Logger'], separator="\t", title='md',
-                                     reportInterval=md_report_interval,
-                                     step=True, totalSteps=totalStepsMD,
-                                     time=False, speed=True, progress=True, remainingTime=True)
-        md_reporters.append(md_progress_reporter)
-        config['simulation']['reporters']['md']['Reporters'] = md_reporters
-
-        #Configure the NCMC simulation reporters
-        ncmc_reporters = []
-        totalStepsNC = config['simulation']['nstepsNC']
-        ncmc_report_interval = config['simulation']['reporters']['ncmc']['reporter_interval']
-        #ncmc_trajectory_interval = config['simulation']['reporters']['ncmc']['trajectory_interval']
-        if 'frame_indices' in config['simulation']['reporters']['ncmc']:
-            frame_indices = config['simulation']['reporters']['ncmc']['frame_indices']
-        else:
-            frame_indices = None
-
-        ncmc_traj_reporter = reporters.NetCDF4Reporter(outfname+'-pmoves.nc', reportInterval=ncmc_report_interval, frame_indices=frame_indices, crds=True, vels=False, frcs=False, protocolWork=True, alchemicalLambda=True)
-        ncmc_progress_reporter = reporters.BLUESStateDataReporter(config['Logger'], reportInterval=ncmc_report_interval,
-                                     separator="\t", title='ncmc',
-                                     step=True, totalSteps=config['simulation']['nstepsNC'],
-                                     time=False, speed=True, progress=True, remainingTime=True)
-
-        ncmc_reporters.append(ncmc_traj_reporter)
-        ncmc_reporters.append(ncmc_progress_reporter)
-        config['simulation']['reporters']['ncmc']['Reporters'] = ncmc_reporters
-        return config
-
-    def set_Parameters(config):
-        """
-        MAIN execution function for updating/correcting (placing units) in the config
-        """
-        try:
-            config = set_Output(config)
-            config = set_Logger(config)
-            config['Structure'] = load_Structure(logger=config['Logger'],**config['structure'])
-            config = set_Units(config)
-            config = set_Apps(config)
-
-            #Check Amber Selections
-            if 'freeze' in config.keys():
-                freeze_keys = ['freeze_center', 'freeze_solvent', 'freeze_selection']
-                for sel in freeze_keys:
-                    if sel in config['freeze']:
-                        check_amber_selection(config['Structure'], config['freeze'][sel], config['Logger'])
-
-            if 'restraints' in config.keys():
-                check_amber_selection(config['Structure'], config['restraints']['selection'], config['Logger'])
-
-            #Calculate NCMC steps with nprop
-            config['simulation']['nstepsNC'], config['simulation']['integration_steps'] = calcNCMCSteps(logger=config['Logger'], **config['simulation'])
-
-            if 'reporters' in config['simulation']:
-                config = set_Reporters(config)
-            else:
-                config['Logger'].error('No simulation reporter parameters found.')
-                sys.exit(1)
-
-        except Exception as e:
-            logger.exception(e)
-            raise
-
-        return config
-
-
-
-    def calcNCMCSteps(total_steps, nprop, prop_lambda, logger, **kwargs):
-        if (total_steps % 2) != 0:
-           logger.exception('`total_steps = %i` must be even for symmetric protocol.' % (total_steps))
-
-        nstepsNC = total_steps/(2*(nprop*prop_lambda+0.5-prop_lambda))
-        if int(nstepsNC) % 2 == 0:
-            nstepsNC = int(nstepsNC)
-        else:
-            nstepsNC = int(nstepsNC) + 1
-
-        in_portion =  (prop_lambda)*nstepsNC
-        out_portion = (0.5-prop_lambda)*nstepsNC
-        if in_portion.is_integer():
-            in_portion= int(in_portion)
-        if out_portion.is_integer():
-            int(out_portion)
-        in_prop = int(nprop*(2*floor(in_portion)))
-        out_prop = int((2*ceil(out_portion)))
-        calc_total = int(in_prop + out_prop)
-        if calc_total != total_steps:
-            logger.warn('total nstepsNC requested ({}) does not divide evenly with the chosen values of prop_lambda and nprop. '.format(total_steps)+
-                           'Instead using {} total propogation steps, '.format(calc_total)+
-                           '({} steps inside `prop_lambda` and {} steps outside `prop_lambda)`.'.format(in_prop, out_prop))
-        logger.info('NCMC protocol will consist of {} lambda switching steps and {} total integration steps'.format(nstepsNC, calc_total))
-        return nstepsNC, calc_total
-
-
-    #Parse YAML into dict
-    if yaml_config.endswith('.yaml'):
-        config = load_yaml(yaml_config)
-
-    #Parse the configions dict
-    if type(config) is dict:
-        config = set_Parameters(config)
-
-    return config
 
 class SystemFactory(object):
     """
@@ -391,7 +59,7 @@ class SystemFactory(object):
         (both sterics and electrostatics components) have to be alchemically
         modified.
     """
-    def __init__(self, structure, atom_indices, **config):
+    def __init__(self, structure, atom_indices, config):
         self.structure = structure
         self.atom_indices = atom_indices
         self.config = config
@@ -704,7 +372,7 @@ class SimulationFactory(object):
         in the NCMC simulation.
     opt : dict of parameters for the simulation (i.e timestep, temperature, etc.)
     """
-    def __init__(self, systems, move_engine, **config):
+    def __init__(self, systems, move_engine, config, md_reporters=None, ncmc_reporters=None):
         #Hide these properties since they exist on the SystemsFactory object
         self._structure = systems.structure
         self._system = systems.md
@@ -713,17 +381,13 @@ class SimulationFactory(object):
         #TODO: change atom_indices selection for multiple regions
         self._atom_indices = move_engine.moves[0].atom_indices
         self.move_engine = move_engine
-
         self.config = config
         self.generateSimulationSet()
 
-        self._md_reporters = config['reporters']['md']['Reporters']
-        for rep in self._md_reporters:
-            self.md.reporters.append(rep)
-
-        self._ncmc_reporters = config['reporters']['ncmc']['Reporters']
-        for rep in self._ncmc_reporters:
-            self.ncmc.reporters.append(rep)
+        if md_reporters:
+            self.md = SimulationFactory.attachReporters(self.md, md_reporters['Reporters'])
+        if ncmc_reporters:
+            self.ncmc = SimulationFactory.attachReporters(self.ncmc, ncmc_reporters['Reporters'])
 
     @classmethod
     def addBarostat(cls, system, temperature=300, pressure=1, frequency=25, **kwargs):
@@ -871,6 +535,12 @@ class SimulationFactory(object):
             val = mmplat.getPropertyValue(simulation.context, prop)
             logger.info('{} = {}'.format(prop,val))
 
+    @staticmethod
+    def attachReporters(simulation, reporters):
+        for rep in reporters:
+            simulation.reporters.append(rep)
+        return simulation
+
     def generateSimulationSet(self):
         """Function used to generate the 3 OpenMM Simulation objects."""
         #Construct MD Integrator and Simulation
@@ -999,10 +669,10 @@ class Simulation(object):
         logger.info('\tTotal MD time = %s ps' % (int(time_md_steps) * int(nIter)))
 
         #Get trajectory frame interval timing for BLUES simulation
-        frame_iter = self.config['nstepsMD'] / self.config['reporters']['md']['trajectory_interval']
-        timetraj_frame = (time_ncmc_steps + time_md_steps) / frame_iter
-        logger.info('\tTrajectory Interval = %s ps' % timetraj_frame)
-        logger.info('\t\t%s frames/iter' % frame_iter )
+        #frame_iter = self.config['nstepsMD'] / self.config['reporters']['md']['trajectory_interval']
+        #timetraj_frame = (time_ncmc_steps + time_md_steps) / frame_iter
+        #logger.info('\tTrajectory Interval = %s ps' % timetraj_frame)
+        #logger.info('\t\t%s frames/iter' % frame_iter )
 
     def getStateInfo(self, context, parameters):
         """Function that gets the State information from the given context and
